+++ src/runtime/src/core/iterator.rs
@@ -119,9 +119,26 @@ pub fn make_module() -> ValueMap {
             }
             Ok(Empty)
         }
+        [iterable, f] if iterable.is_iterable() && f.is_callable() => {
+            let f = f.clone();
+            for output in make_iterator(iterable).unwrap() {
+                let run_result = match output {
+                    Output::Value(value) => vm.run_function(f.clone(), CallArgs::Single(value)),
+                    Output::ValuePair(a, b) => {
+                        vm.run_function(f.clone(), CallArgs::AsTuple(&[a, b]))
+                    }
+                    Output::Error(error) => return Err(error),
+                };
+
+                if run_result.is_err() {
+                    return run_result;
+                }
+            }
+            Ok(Empty)
+        }
         unexpected => unexpected_type_error_with_slice(
             "iterator.consume",
-            "an iterable value as argument",
+            "an Iterable Value (and optional Function) as arguments",
             unexpected,
         ),
     });